+++ src/main/java/com/github/johnmedlockdev/main/Main.java
@@ -4,67 +4,27 @@ import com.github.johnmedlockdev.main.data.Data;
 import com.github.johnmedlockdev.main.file.FileInfo;
 import com.github.johnmedlockdev.main.parse.ParseInput;
 
-import java.io.*;
-import java.util.Random;
+import java.io.IOException;
 
-class Main {
-    public static void main(String ...args) throws IOException {
+public class Main {
+    public static void main(String... args) throws IOException {
 
         ParseInput userInput = new ParseInput(args);
         FileInfo fileInfo = new FileInfo(userInput);
+        Data data = new Data(userInput, fileInfo);
 
-        // create instance
-        Data asset = new Data(userInput.getTicker(), userInput.getPrice(), fileInfo.getFileFullName());
-
-        // control flow for based off user operation selection
-        if (userInput.getMethod().equals("input")) {
-            asset.setPrice(userInput.getPrice());
+        if (data.getMethod().equals("new")) {
+            data.createFile();
+        } else if (data.getMethod().equals("add")) {
+            data.createInput();
+        } else if (data.getMethod().equals("generate")) {
+            data.generate();
+        } else if (data.getMethod().equals("predict")) {
+            data.getPrediction();
+        }else {
+            System.out.println("Not a valid method.");
         }
 
-        if (userInput.getMethod().equals("predict")) {
-            System.out.println(asset.getPrediction());
-        }
-
-        /// ------------------------------------------------
-
-        // need to figure out how to make paths dynamic
-
-        if (fileInfo.getPath().exists() && userInput.getMethod().equals("input")) {
-            String contentsToWrite = Double.toString(userInput.getPrice());
-            BufferedWriter writer = new BufferedWriter(
-                    new FileWriter(fileInfo.getFileFullName(), true)
-            );
-            writer.newLine();   //Add new line
-            writer.write(contentsToWrite);
-            writer.close();
-
-            System.out.println("The data file exist");
-        } else if (!fileInfo.getPath().exists() && userInput.getMethod().equals("input")) {
-            OutputStream outputStream = new FileOutputStream(fileInfo.getPath());
-            String contentsToWrite = Double.toString(userInput.getPrice());
-            outputStream.write(contentsToWrite.getBytes());
-            outputStream.close();
-            System.out.println("New file created for " + userInput.getTicker());
-        }
-
-
-        if (fileInfo.getPath().exists() && userInput.getMethod().equals("generate")) {
-            String contentsToWrite = Double.toString(userInput.getPrice());
-            for (int i = 0; i < 200; i++) {
-                Random r = new Random(); // creating Random object
-                double randomValue = 1 + (10 - 1) * r.nextDouble();
-                String ranStr = String.format("%.2f", randomValue);
-
-                BufferedWriter writer = new BufferedWriter(
-                        new FileWriter(fileInfo.getFileFullName(), true)
-                );
-                writer.newLine();   //Add new line
-                writer.write(ranStr);
-                writer.close();
-            }
-
-
-        }
 
     }
 }